@@ -11,7 +11,7 @@ maintainers = [{ name = "John Aziz", email = "johnaziz269@gmail.com" }]
 license = {file = "LICENSE"}
 readme = "README.md"
 
-requires-python = ">=3.8"
+requires-python = ">=3.9"
 dependencies = [
     "requests",
     "types-requests",
@@ -21,7 +21,6 @@ classifiers = [
     "License :: OSI Approved :: MIT License",
     "Operating System :: OS Independent",
     "Programming Language :: Python :: 3",
-    "Programming Language :: Python :: 3.8",
     "Programming Language :: Python :: 3.9",
     "Programming Language :: Python :: 3.10",
     "Programming Language :: Python :: 3.11",
@@ -40,26 +39,29 @@ Releases = "https://github.com/john0isaac/markdown-checker/releases"
 
 [project.optional-dependencies]
 dev = [
-    "pytest",
-    "pytest-cov",
     "pre-commit",
-    "ruff",
-    "black",
-    "mypy",
+    "mypy"
+]
+lint = [
+    "ruff"
+]
+test = [
+    "pytest",
+    "pytest-cov"
+]
+docs = [
+    "mkdocs",
+    "mkdocs-material",
+    "mkdocstrings[python]",
+    "pymdown-extensions"
 ]
 
 [tool.ruff]
+lint.select = ["E", "F", "I", "UP"]
 line-length = 120
-target-version = "py38"
+target-version = "py39"
 src = ["src"]
 output-format = "full"
-
-[tool.ruff.lint]
-select = ["E", "F", "I", "UP"]
-
-[tool.black]
-line-length = 120
-target-version = ["py38"]
 
 [project.scripts]
 markdown-checker = "markdown_checker.validate_markdown:main"
